Simplify use of primary and foreign keys
This only works for single-column primary keys, but
allows the key to be specified by name in the keyword
arguments of the route. The named key is converted to
KEY__pk for filtering the model's queryset.

Foreign keys may be specified in POST data using the
foreign key name. Once the new model instance has
been created, the keys will be assigned via KEY_id.